requirements.txt
click
python-twitter
